+++ src/capi/azext_capi/helpers/network_helpers.py
